Split element info into separate component
